# crates/quest/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CMU Orientation Quest",
    "description": "Backend for the Orientation Quest app.",
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/passes/apple/challenge": {
      "get": {
        "tags": ["passes"],
        "operationId": "challenge",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PassChallenge"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrBody"
                }
              }
            }
          },
          "502": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/staff/card": {
      "post": {
        "tags": ["staff"],
        "operationId": "read",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardUrl"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardView"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrBody"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrBody"
                }
              }
            }
          },
          "502": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/staff/card/{card_id}/challenge": {
      "put": {
        "tags": ["staff"],
        "operationId": "link",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "description": "Uppercase hex card UID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardView"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrBody"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrBody"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrBody"
                }
              }
            }
          },
          "502": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrBody"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["staff"],
        "operationId": "unlink",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "description": "Uppercase hex card UID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardView"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrBody"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrBody"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrBody"
                }
              }
            }
          },
          "502": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/staff/card/{card_id}/location": {
      "put": {
        "tags": ["staff"],
        "operationId": "place",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "description": "Uppercase hex card UID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardView"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrBody"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrBody"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrBody"
                }
              }
            }
          },
          "502": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/me": {
      "get": {
        "tags": ["users"],
        "operationId": "me",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrBody"
                }
              }
            }
          },
          "502": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/me/dorm": {
      "put": {
        "tags": ["users"],
        "operationId": "dorm",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DormBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DormBody"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrBody"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrBody"
                }
              }
            }
          },
          "502": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/challenge": {
      "get": {
        "tags": ["devices"],
        "operationId": "challenge",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Challenge"
                }
              }
            }
          },
          "502": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrBody"
                }
              }
            }
          }
        },
        "security": [{}]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuthErrBody": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "CardUrl": {
        "type": "object",
        "required": ["url"],
        "properties": {
          "url": {
            "type": "string"
          }
        }
      },
      "CardView": {
        "type": "object",
        "required": ["card_id"],
        "properties": {
          "card_id": {
            "type": "string"
          },
          "challenge_id": {
            "type": ["string", "null"],
            "format": "uuid"
          },
          "lat": {
            "type": ["number", "null"],
            "format": "double"
          },
          "lon": {
            "type": ["number", "null"],
            "format": "double"
          }
        }
      },
      "Challenge": {
        "type": "object",
        "required": ["nonce", "expires_in"],
        "properties": {
          "expires_in": {
            "type": "integer",
            "format": "int64"
          },
          "nonce": {
            "type": "string"
          }
        }
      },
...
      "DormBody": {
        "type": "object",
        "required": ["dorm"],
        "properties": {
          "dorm": {
            "type": "string"
          }
        }
      },
...
      "LinkBody": {
        "type": "object",
        "required": ["challenge_id"],
        "properties": {
          "challenge_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "PassChallenge": {
        "type": "object",
        "required": ["andrew_id", "issued_at", "message"],
        "properties": {
          "andrew_id": {
            "type": "string"
          },
          "issued_at": {
            "type": "integer",
            "format": "int64"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "PlaceBody": {
        "type": "object",
        "required": ["lat", "lon"],
        "properties": {
          "lat": {
            "type": "number",
            "format": "double"
          },
          "lon": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "Profile": {
        "type": "object",
        "required": ["andrew_id", "created_at"],
        "properties": {
          "andrew_id": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "dorm": {
            "type": ["string", "null"]
          }
        }
      },
...
    },
    "securitySchemes": {
      "device_proof": {
        "type": "apiKey",
        "in": "header",
        "name": "x-device-proof",
        "description": "Signature over the request method and URL by an enrolled device key"
      },
      "session": {
        "type": "http",
        "scheme": "bearer",
        "description": "Session id issued by /auth/login"
      }
    }
  },
  "security": [
    {
      "device_proof": [],
      "session": []
    }
  ],
...
}